requirements.txt
connexion[flask,swagger-ui]==2.14.2
uvicorn==0.23.2
Flask-SocketIO==5.2.0
python-decouple==3.8
torch==2.1.0
uuid6==2023.5.2
numpy==1.26.1